Organized errors reporting function
Separated the error, error message, and error log classes into separated files for better organization. Replaced repetitive code using newly created functions.
