+++ mirri/io/writers/error_logging/__init__.py
@@ -1,0 +1,3 @@
+from .error import Entity, Error
+from .error_message import ErrorMessage
+from .error_log import ErrorLog
